pyproject.toml
[tool.poetry]
name = "runner-project"
version = "0.1.0"
description = "This project ia webpage designed for runners to track their progress, share routes, and connect with the running community."
authors = ["DD"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.1"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
